backend/oauth_storage/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client